adding routes to mock server
