+++ modules/mock-server/src/main/scala/com/lumidion/sonatype/central/mockserver/DeploymentRepository.scala
@@ -1,6 +1,11 @@
 package com.lumidion.sonatype.central.mockserver
 
 import com.lumidion.sonatype.central.client.core.{DeploymentName, DeploymentState}
+import com.lumidion.sonatype.central.mockserver.router.Error.{
+  DeploymentNotFound,
+  InvalidPublishableDeployment,
+  PublishDeploymentError
+}
 
 import java.util.UUID
 import scala.collection.mutable.{HashMap => MutableHashMap}
@@ -23,6 +28,25 @@ class DeploymentRepository {
     for {
       deployment <- getDeployment(id)
       res = mutableMap.update(id, (deployment._1, state))
+    } yield res
+  }
+
+  def deleteDeployment(id: UUID): Option[Unit] = {
+    mutableMap.get(id).map { _ =>
+      mutableMap -= id
+    }
+  }
+
+  def publishValidatedDeployment(id: UUID): Either[PublishDeploymentError, Unit] = {
+    for {
+      deployment <- mutableMap.get(id).toRight(DeploymentNotFound)
+      res <-
+        if (deployment._2 == DeploymentState.VALIDATED) {
+          mutableMap.update(id, (deployment._1, DeploymentState.PUBLISHED))
+          Right(())
+        } else {
+          Left(InvalidPublishableDeployment)
+        }
     } yield res
   }
 }
